+++ src/utils/utils.ts
@@ -1,7 +1,7 @@
 import BigNumber from "bignumber.js";
 import { INTERNAL_ROUTES, QUERY_PARAMS, THEMES } from "../constants/constants";
+import { Asset, ChainAsset, Swap } from "@gardenfi/orderbook";
 import { Assets } from "../store/assetInfoStore";
-import { Asset, Swap } from "@gardenfi/orderbook";
 
 export const isProduction = () => {
   return import.meta.env.VITE_ENVIRONMENT === "production";
@@ -27,7 +27,12 @@ export const capitalizeChain = (chainKey: string) => {
  * @returns
  */
 export const getAssetFromSwap = (swap: Swap, assets: Assets | null) => {
-  return assets && assets[`${swap.chain}_${swap.asset.toLowerCase()}`];
+  if (!assets) return;
+  return Object.values(assets).find(
+    (asset) =>
+      ChainAsset.from(asset.id).toString() ===
+      ChainAsset.from(swap.asset).toString()
+  );
 };
 
 export const getQueryParams = (urlParams: URLSearchParams) => {
@@ -93,9 +98,10 @@ export const isCurrentRoute = (route: string) => {
 };
 
 export const formatAmountUsd = (
-  amount: string | number | bigint,
+  amount: string | number | bigint | undefined,
   decimals: number
 ) => {
+  if (!amount) return 0;
   const num = formatAmount(amount, decimals);
   return Number(num).toLocaleString("en-US", {
     maximumFractionDigits: 2,
@@ -151,7 +157,7 @@ export const getFirstAssetFromChain = (
 
   const assetKey = Object.keys(assets).find((key) => {
     const asset = assets[key];
-    return asset.chain === chain && !asset.disabled;
+    return asset.chain === chain;
   });
 
   return assetKey ? assets[assetKey] : undefined;