+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## [1.1.8]
 
+### Added
+
+- Added starknet support for Xverse wallet. 
+
 ### Fixed
 
 - Update toast for successful swap and add abort controller to fetch assets function.
 - Update the coinbase wallet id.
 - Fixed the link redirecting to staking docs.
+- Disable Core chain in output-asset on Phantom, Keplr and Leap wallet.
 
 ## [1.1.7]
 
@@ -28,6 +33,7 @@
 - Updated balance retrieval function to use fallback RPC clients as default instead of default wallet RPCs.
 - Rate conversion and display estimated token price for inputAsset.
 - Fixed fiat token prices.
+- Add Tokeo wallet support for SUI Mainnet
 
 ## [1.1.4]